% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BT_ASR_matrix.R
\name{BT_ASR_matrix}
\alias{BT_ASR_matrix}
\title{This function takes a BayesTraits output. The matrix sould be either provided as a data frame()see BT_summarize function or read using this function from a given foldder}
\usage{
BT_ASR_matrix(ASR_matrix = NULL, nstates = 2, bn = 0.1, read = T,
  folder = ".")
}
\arguments{
\item{ASR_matrix}{should have the following structure RecNodeX.P.1. and RecNodeX.P.0.}

\item{nstates}{Number of discrete states in the analysis}

\item{bn}{burnin (0-1)}

\item{read}{\code{logical} whether results should be read from the selected \code{folder}}

\item{folder}{where the output files of a single BayesTraits analysis are found}
}
\description{
This function takes a BayesTraits output. The matrix sould be either provided as a data frame()see BT_summarize function or read using this function from a given foldder
}
