% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BT_multiple.R
\name{BT_multiple}
\alias{BT_multiple}
\title{Input files for multiple BT runs}
\usage{
BT_multiple(tree, dataset, tip.label, trait, run = F, name_general,
  ML = F, it = 1.1e+07, bur = 1e+06)
}
\arguments{
\item{tree}{A fully-bifurcating Phy object}

\item{dataset}{A data.frame object with two columns (tip name + state[numeric])}

\item{tip.label}{A \code{character} indicating the name of the column with the trait}

\item{trait}{A \code{character} indicating the name of the column with the tip names in the dataset}

\item{run}{\code{logic} whether analyses should start after input files are created}

\item{name_general}{A \code{character} indicating the name of the analuysis (use unique per analysis)}

\item{ML}{\code{logic} whether parameters should be optimized under a Maximum Likelihood approach. MCMC analyses are available under \code{ML=F}}

\item{it}{The number of iterations under a bayesian optimization (only when \code{ML=F})}

\item{bur}{Total number of generations assumed as burnin under a bayesian optimization (only when \code{ML=F})}
}
\description{
This function creates imput files for a large combination of the parameters
 under the BayesTraits MultiState model.
}
\details{
Specifically, this function is designed to run models differing in the following
  parameters: (1) "PriorAll" vs. "RevJump", (2) Variable rates vs. non-Variable rates,
 and (3) uniform vs. exponential distributions.
}
