% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BT_progress.R
\name{BT_progress}
\alias{BT_progress}
\title{Progress bar for multiple BT analyses}
\usage{
BT_progress(...)
}
\arguments{
\item{...}{Not requiered}
}
\description{
A progress bar for BayesTraits analyses
}
