% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BT_summarize.R
\name{BT_summarize}
\alias{BT_summarize}
\title{Summary of BT analyses}
\usage{
BT_summarize(folder = ".", bn = 0)
}
\arguments{
\item{folder}{where the output files of a BayesTraits analysis are found}

\item{bn}{burnin (0-1)}
}
\description{
Burnin may have been applied in a previous stage of the analysis
}
