% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circular.time.slice.R
\name{circular.time.slice}
\alias{circular.time.slice}
\title{mcc2}
\usage{
circular.time.slice(phy, start, end, col = "grey")
}
\arguments{
\item{phy}{\code{phylo} tre object}

\item{start}{start of the time slice (lower)}

\item{end}{end of the time slice (upper)}

\item{col}{choose color (default = "grey")}
}
\description{
Adds a circular time slice to a ploted \code{phylo} tree
of \code{type = "fan"}
}
\examples{
path <- system.file("data/mcc.tre", package="rBt")
tr <- read.beast.annot(path)
timerange <- range(branching.times(tr))
# equally spaced time slices
timeslices <- seq(timerange[1], timerange[2], length.out=5)
ts <- cbind(timeslices[-5], timeslices[-5]+((timeslices[3]-timeslices[2])/2))
# plot tree
plot(tr, type="fan", show.tip.label=F)
# add slice
for (i in 1:dim(ts)[1])
   circular.time.slice(tr, ts[i,1], ts[i,2])
}
\seealso{
\code{\link{plot.phylo}} \code{\link{sp}}
}
