% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\docType{data}
\name{dataset}
\alias{dataset}
\title{Bayestraits sample data. Artiodactyla}
\format{
A data frame with 17 species and a single character:
}
\source{
\url{http://www.NNN.info/}
}
\usage{
dataset
}
\description{
Artiodactyla sample data from Bayestraits
}
\keyword{datasets}
