% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.tip.spp.R
\name{get.tip.spp}
\alias{get.tip.spp}
\title{get.tip.spp}
\usage{
get.tip.spp(phy, del = "_")
}
\arguments{
\item{phy}{a \code{phylo} object}

\item{del}{a \code{character} string with the delimitator.
Default is \code{"_"}.}
}
\value{
a \code{character} vector
}
\description{
Returns a vector is species names based on a \code{phylo}  
object's tip names.
}
\examples{
file <- system.file("data/mcc.tre", package="rBt")
tr <- read.nexus(file)
spp <- get.tip.spp(tr, del="__")
tipcols <- rainbow(8)[ factor(spp) ]
}
