% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsi.R
\name{gsi}
\alias{gsi}
\title{gsi}
\usage{
gsi(phy, groups)
}
\arguments{
\item{phy}{a \code{phylo} object}

\item{groups}{a \code{list} where each element is a character
vector with groups of tips}
}
\value{
a \code{vector} with gsi values
}
\description{
Computes genealogical sorting index (Cummings et al. 2008)
}
\examples{
set.seed(1)
tr = rtree(10)
groups = split(tr$tip.label, factor(rep(c(1,2,3), c(3,3,4))))
tr_gsi = gsi(tr, groups)
tr_gsi
#     1     2     3 
# 1.000 1.000 0.625
}
\references{
Cummings, M.P., M.C. Neal, K L. Shaw (2008) A genealogical approach to 
            quantifying lineage divergence. Evolution 62:2411-2422
            \url{doi:10.1111/j.1558-5646.2008.00442.x}
}
