% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ltt.df.hpd.R
\name{ltt.df.hpd}
\alias{ltt.df.hpd}
\title{ltt.df.hpd}
\usage{
ltt.df.hpd(phy)
}
\arguments{
\item{phy}{a \code{phylo} object read by \code{read.annot.beast}}
}
\value{
\code{list} object with 2 data frames, one for points and the other for polygons
}
\description{
Produces a list with two data frames; one with
mean node values that can be represented as points
and or lines, and a another with 95% HPD values that
that can be represented as a polygon.
}
\details{
It's inteded use is to generate a format that is
easily compatible with ggplot2 (see example).
}
\examples{
file <- system.file("data/mcc.tre", package="rBt")
tr <- read.annot.beast(file)
# find the root node:
my_ltt <- ltt.df.hpd(tr)
library(ggplot2)
ggplot(my_ltt$points, aes(x=mean, y=N)) +
    geom_poly(data=my_ltt$poly, aes(x=x, y=y), fill="grey", alpha=0.5) +
    geom_point() + geom_line()
}
