% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcc.R
\name{mcc}
\alias{mcc}
\title{mcc}
\usage{
mcc(phy)
}
\arguments{
\item{phy}{List of \code{multiPhylo} trees}
}
\value{
mcc tree \code{phylo} object
}
\description{
This function is based on the \code{\link{maxCladeCred}}
in \code{phangorn}, and it finds the maximum clade 
crebility tree given a list of trees.
}
\examples{
path <- system.file("data/trees/", package="rBt")
trs <- mcc.trees2multi(path)
mcctr <- mcc(trs)





}
\seealso{
\code{\link{maxCladeCred}} \code{\link{prop.part}}
}
