% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcc.trees2multi.R
\name{mcc.trees2multi}
\alias{mcc.trees2multi}
\title{mcc.trees2multi}
\usage{
mcc.trees2multi(path = ".")
}
\arguments{
\item{path}{The path to the directory with the MCC files.
By default it will list files in the current path(.).}
}
\value{
list of trees of class \code{multiPhylo}
}
\description{
This function reads separate MCC files from BEAST
an stores them as a \code{multiPhylo} object.
}
\examples{
path <- system.file("data/trees/", package="rBt")
trs <- mcc.trees2multi(path)
class(trs)
# [1] "multiPhylo"
}
\seealso{
\code{\link{read.beast.annot}} \code{\link{read.nexus}}
}
