% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/order.edges.R
\name{order.edges}
\alias{order.edges}
\title{order.edges}
\usage{
order.edges(phy)
}
\arguments{
\item{phy}{tree, \code{phylo} object}
}
\value{
\code{numeric} vector with edge indices
}
\description{
Gets the edge index ordered by node (e.g.
order by node index). This function is practical
for plotting node values on edges instead of nodes.
}
\note{
tree files read by \code{read.beast.annot} will include an
      element with ordered edges: phy$edge.ordered
}
\examples{
file <- system.file("data/trees/mcc.tre", package="rBt")
tr <- read.nexus(tr)
nodes <- (length(tr$tip.label)+1):(length(tr$tip.label)+tr$Nnode)
edges.ordered <- order.edges(tr)
plot(tr)
edgelabels(edge=edges.ordered, text=nodes[-1]) # -1 is needed to exclude the root node
}
\seealso{
\code{\link{edgelabels}} \code{\link{read.beast.annot}}
}
