% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylogeoToRasterHPD.R
\name{phylogeoToRasterHPD}
\alias{phylogeoToRasterHPD}
\title{gphylogeoToRasterHPD}
\usage{
phylogeoToRasterHPD(trees, coordnames = NULL, intervals = NULL)
}
\arguments{
\item{file}{The path to the MCC file from BEAST}
}
\value{
\code{raster} object
}
\description{
Extracts coordinate data from a full posterior BEAST trees file
read by \code{read.annot.beast} and generates \code{raster} objects
with 2d kernel HPD densities.
}
\details{
Thee trees object must be \code{multiPhylo} or \code{list}.
It requires two additional arguments: \code{coordnames}, which is 
character vector with the column names pointing to the coordinate 
data in the BEAST tree (see examples); and \code{intervals}, which is
a numeric vector with the time intervals or slices that will be used 
to summarize and claculate kernel densities.
}
\examples{
file <- system.file("data/HBV.trees", package="rBt")
trees <- read.annot.beast(file)
# remove burnin
trees <- trees[1000:10000]
# look for column names
head(trees[[1]]$metadata)
# node            locationsnewTrait
#1    1 33.3431887795,134.9307404236
#2    2 -2.3045941592,118.5275669247
#3    3  15.880672531,106.7863444469
#4    4  15.2357758333,97.1170427803
#5    5  10.826174473,100.0321841486
#6    6 11.8647366524,104.4581350327
coordnames <- "locationsnewTrait"
# set intervals
intervals <- c(0,250,500,750)
# get 95\% HPD raster
st <- phylogeoToRasterHPD(trees, coordnames, intervals)
plot(st)
# add a map:
library(maptools)
data(wrld_simpl)
for (i in 1:dim(st)[3]){
   image(st[[i]], col=rev(terrain.colors(20)))
   plot(wrld_simpl, add=T)
}
}
\seealso{
\code{\link{raster}}
}
