% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.phylo.HPD.R
\name{plot.phylo.HPD}
\alias{plot.phylo.HPD}
\title{plot.phylo.HPD}
\usage{
plot.phylo.HPD(x, nodes = NULL, pb = FALSE, bar.width = 0.3,
  bar.col = NA, border = NULL, at = NULL, minor = NULL,
  vline = TRUE, ...)
}
\arguments{
\item{x}{annotated \code{phylo} object}

\item{nodes}{a vector with node numbers to 
plot. Default is \code{NULL}}

\item{pb}{TRUE if the tree is a chronogram from PhyloBayes with HPD}

\item{bar.width}{fraction of 1 for bar thickness}

\item{bar.col}{bar color}

\item{border}{if bar borders should be drawn. See ?rect.}

\item{vline}{add a vertical lines (default: TRUE)}

\item{...}{further arguments passed by \code{phylo.plot}}
}
\value{
plot
}
\description{
This function plots a \code{phylo} BEAST tree read by
\code{read.annot.beast} with its 95% HPD node height
intervals and adds a x-axis scale.
}
\details{
** Now includes support for PhyloBayes chronograms **
}
\examples{
file <- system.file("data/mcc.tre", package="rBt")
tr <- read.annot.beast(file)
# plot HPD heights on all nodes:
plot.phylo.HPD(tr, cex=0.5, bar.width=0.4, bar.col="red", border=NA)
# with library(scales) for transparency
# plot.phylo.HPD(tr, cex=0.5, bar.width=0.4, bar.col=alpha("red",0.5), border=NA)
# plot HPD heights on specific nodes:
tips <- tr$tip.label
tipsfac <- factor(sapply(tips, function(x) strsplit(x, "__")[[1]][1] ))
names(tipsfac) <- NULL
tipslist <- split(tips, tipsfac)
allanc <- c(unlist(sapply(tipslist, function(x, y=tr) Ancestors(tr, which.node(y, x)) )), 
            sapply(tipslist, function(x) which.node(tr, x)))
allanc <- sort(allanc)
allanc <- allanc[ !duplicated(allanc) ]
names(allanc) <- NULL
plot.phylo.HPD(tr, cex=0.5, bar.width=0.4, bar.col="red", border=NA, nodes=allanc)
}
\seealso{
\code{\link{plot.phylo}} \code{\link{rect}}
}
