% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.annot.beast.R
\name{read.annot.beast}
\alias{read.annot.beast}
\title{rread.annot.beast}
\usage{
read.annot.beast(file)
}
\arguments{
\item{file}{The path to the MCC file from BEAST}
}
\value{
annotated \code{phylo} object
}
\description{
This function reads a nexus formated beast file and 
appends all meta-data and annotations.
}
\details{
Node ordering is the same as in the \code{read.tree}/\code{read.nexus}
functions from ape (cladewise). The metadata is stored as a 
data.frame named 'metadata'. Only node posterior
probabilites are passed as an additional numeric vector
named 'posterior'. Note that 'metadata' includes tip 
annotations as well. The code might be a bit slow for large trees.
}
\examples{
file <- system.file("data/mcc.tre", package="rBt")
tr <- read.annot.beast(file)
class(tr)
# [1] "phylo"
names(tr)
#[1] "edge"        "edge.length" "Nnode"       "root.edge"   "tip.label"  
#[6] "metadata"    "posterior"
# for add pp to nodes you could try:
plot(tr)
nodelabels(text=round(tr$posterior,2), cex=0.5)
# for edges, try:
tr$ordered.edges <- order.edges(tr)
plot(tr)
edgelabels(edge=tr$ordered.edges, text=round(tr$posterior,2), cex=0.5)
}
\seealso{
\code{\link{read.nexus}}
}
