% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.beast.annot.R
\name{read.beast.annot}
\alias{read.beast.annot}
\title{read.beast.annot}
\usage{
read.beast.annot(file)
}
\arguments{
\item{file}{The path to the MCC file from BEAST}
}
\value{
annotated \code{phylo} object
}
\description{
This function reads a nexus formated MCC beast file and 
appends all meta-data and annotations. It uses a Perl 
script internally that quickly generates a table with 
annotations. Node ordering is the same as in the 
read.tree/read.nexus function from ape (cladewise).
}
\examples{
file <- system.file("data/mcc.tre", package="rBt")
tr <- read.beast.annot(file)
class(tr)
# [1] "phylo"
names(tr)
# [1] "edge"                   "edge.length"            "Nnode"                 
# [4] "root.edge"              "tip.label"              "nodes"                 
# [7] "CAheight_mean"          "CAheight_median"        "height"                
#[10] "height_median"          "length"                 "length_median"         
#[13] "posterior"              "CAheight_95\%_HPD_LOWER" "CAheight_95\%_HPD_UPPER"
#[16] "height_95\%_HPD_LOWER"   "height_95\%_HPD_UPPER"   "length_95\%_HPD_LOWER"  
#[19] "length_95\%_HPD_UPPER"   "edge.ordered"           "data"
}
\seealso{
\code{\link{read.nexus}}
}
