% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RplotEBS.R
\name{RplotEBS}
\alias{RplotEBS}
\title{RplotEBS}
\usage{
RplotEBS(file = NULL, path = ".", pattern = ".csv", trim.x = NULL,
  trim.y = NULL, ...)
}
\arguments{
\item{path}{a \code{character} string with the path to 
CSV files (\code{character}). Default is ".".}

\item{pattern}{a \code{character} string pattern to search 
for CSV files in the \code{path}. Default is ".csv".}

\item{trim.x}{a \code{numeric} value indicating the lower limit
to the x-axis.}

\item{trim.y}{a \code{numeric} value indicating the upper limit
to the y-axis.}
}
\value{
a \code{plot}
}
\description{
Plots one or multiple EBSP given a CSV ouput from BEAST.
}
\examples{
file <- system.file("data/EBSP.csv", package="rBt")
RplotEBS(file)
}
