% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpleAxisPhylo.R
\name{simpleAxisPhylo}
\alias{simpleAxisPhylo}
\title{simpleAxisPhylo}
\usage{
simpleAxisPhylo(at = NULL, minor = NULL, ...)
}
\arguments{
\item{at}{a vector of points where to place major tickmarks}

\item{minor}{NULL or numeric factor by which to divide intervals}

\item{...}{further arguments passed to \code{axis}}
}
\value{
x-axis (side 1)
}
\description{
Adds an x-axis to the latest plotted tree with minor
with custom tickmarks and minor tickmarks.
}
\examples{

# simulate coalescent tree and plot
set.seed(1)
tr <- rcoal(10)
plot(tr)

# add default axis
simpleAxisPhylo()

# add custom scale
plot(tr)
trange <- c(0, max(branching.times(tr)))
trange <- pretty(trange)
x <- seq(trange[1], rev(trange)[1], (diff(trange)/2)[1])
simpleAxisPhylo(at=x, minor=2)
}
\seealso{
\code{\link{axis}} \code{\link{axisPhylo}}
}
