% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree.R
\docType{data}
\name{tree}
\alias{tree}
\title{Bayestraits sample tree Artiodactyla}
\format{
A data frame with 17 species:
}
\source{
\url{http://www.NNN.info/}
}
\usage{
tree
}
\description{
Artiodactyla sample tree from Bayestraits
}
\keyword{datasets}
