% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/which.tips.R
\name{which.tips}
\alias{which.tips}
\title{which.tips}
\usage{
which.tips(phy, node, text = FALSE)
}
\arguments{
\item{phy}{a \code{phylo} object}

\item{node}{a \code{vector} corresponding to a single node
in \code{numeric} format}

\item{text}{a \code{boolean} indicating if tips are to be
returned as text (default: FALSE)}
}
\value{
\code{numeric} vector object with tips
}
\description{
Finds a set of tips associated to a given node.
}
\examples{
file <- system.file("data/mcc.tre", package="rBt")
tr <- read.nexus(file)
# find the root node:
tips_in_node <- which.tips(tr, 130)
}
