% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BT_single.R
\name{BT_single}
\alias{BT_single}
\title{Input files for a single BT run}
\usage{
BT_single(tree, dataset, tip.label, trait, ML = ML, model = "RevJump",
  vrates = T, Iterations = 10000, Burnin = 100, dist = "exp",
  val_prior = 10, name = "Default", run = F)
}
\arguments{
\item{tree}{A fully-bifurcating Phy object}

\item{dataset}{A data.frame object with two columns (tip name + state[numeric])}

\item{tip.label}{A \code{character} indicating the name of the column with the trait}

\item{trait}{A \code{character} indicating the name of the column with the tip names in the dataset}

\item{ML}{\code{logic} whether parameters should be optimized under a Maximum Likelihood approach. MCMC analyses are available under \code{ML=F}}

\item{model}{A \code{character} vector used to set the analysis to a "PriorAll" or "RevJump" model}

\item{vrates}{Whether a VarRates model should be abalyzed, \code{logic}}

\item{Iterations}{The number of iterations under a bayesian optimization (only when \code{ML=F})}

\item{Burnin}{Total number of generations assumed as burnin under a bayesian optimization (only when \code{ML=F})}

\item{dist}{A \code{character} vector indicating whether the distribution is "exp" (exponential) or "uniform" (uniform)}

\item{val_prior}{A vector of class \code{numeric} with the parameters for the "exp" (1 parameter) or "uniform (2 parameters) distribution}

\item{name}{The name of the run. This name is used to create a subfolder within the working directory}

\item{run}{\code{logic} whether analyses should start after input files are created}
}
\description{
This function creates all input files for a single analysis under the BayesTraits
MultiState model.
}
